# Optimize static codec serialization