--- a/src/Orleans.CodeGenerator/Model/ICodecDescription.cs
+++ b/src/Orleans.CodeGenerator/Model/ICodecDescription.cs
@@ -2,11 +2,6 @@ using Microsoft.CodeAnalysis;
 
 namespace Orleans.CodeGenerator
 {
-    internal interface ICodecDescription
-    {
-        ITypeSymbol UnderlyingType { get; }
-    }
-
     internal interface ICopierDescription
     {
         ITypeSymbol UnderlyingType { get; }