--- a/src/Orleans.CodeGenerator/SerializerGenerator.cs
+++ b/src/Orleans.CodeGenerator/SerializerGenerator.cs
@@ -410,6 +410,10 @@ namespace Orleans.CodeGenerator
                     codecExpression = IdentifierName(instanceCodec.FieldName);
                 }
 
+               // When a static codec is available, we can call it directly and can skip passing the expected type, since it is known to be the static codec's field type:
+               //   C#: <staticCodec>.WriteField(ref writer, <fieldIdDelta, <member>)
+               // When no static codec is available:
+               //   C#: <codecField>.WriteField(ref writer, <fieldIdDelta>, <expectedType>, <member>)
                 var writeFieldArgs = new List<ArgumentSyntax> {
                     Argument(writerParam).WithRefOrOutKeyword(Token(SyntaxKind.RefKeyword)),
                     Argument(fieldIdDeltaExpr)