--- a/src/Orleans.CodeGenerator/SerializerGenerator.cs
+++ b/src/Orleans.CodeGenerator/SerializerGenerator.cs
@@ -410,7 +410,8 @@ namespace Orleans.CodeGenerator
                     codecExpression = IdentifierName(instanceCodec.FieldName);
                 }
 
-               // When a static codec is available, we can call it directly and can skip passing the expected type, since it is known to be the static codec's field type:
+               // When a static codec is available, we can call it directly and can skip passing the expected type,
+               // since it is known to be the static codec's field type:
                //   C#: <staticCodec>.WriteField(ref writer, <fieldIdDelta, <member>)
                // When no static codec is available:
                //   C#: <codecField>.WriteField(ref writer, <fieldIdDelta>, <expectedType>, <member>)