… src/Orleans.Serialization/Buffers/Writer.cs
… src/Orleans.Serialization/Codecs/ByteArrayCodec.cs
… src/Orleans.Serialization/Codecs/DateOnlyCodec.cs
… src/Orleans.Serialization/Codecs/DateTimeCodec.cs
… src/Orleans.Serialization/Codecs/DateTimeOffsetCodec.cs
… src/Orleans.Serialization/Codecs/DictionaryCodec.cs
… src/Orleans.Serialization/Codecs/Enum32BaseCodec.cs
… src/Orleans.Serialization/Codecs/FloatCodec.cs
… src/Orleans.Serialization/Codecs/GuidCodec.cs
… src/Orleans.Serialization/Codecs/IPAddressCodec.cs
… src/Orleans.Serialization/Codecs/IPEndPointCodec.cs
… src/Orleans.Serialization/Codecs/IntegerCodec.cs
… src/Orleans.Serialization/Codecs/ListCodec.cs
… src/Orleans.Serialization/Codecs/ObjectCodec.cs
… src/Orleans.Serialization/Codecs/QueueCodec.cs
… src/Orleans.Serialization/Codecs/ReferenceCodec.cs
… src/Orleans.Serialization/Codecs/StackCodec.cs
… src/Orleans.Serialization/Codecs/StringCodec.cs
… src/Orleans.Serialization/Codecs/TimeOnlyCodec.cs
… src/Orleans.Serialization/Codecs/TimeSpanCodec.cs
… src/Orleans.Serialization/Codecs/TypeSerializerCodec.cs
… src/Orleans.Serialization/Codecs/UriCodec.cs
… src/Orleans.Serialization/Codecs/ValueTupleCodec.cs
… src/Orleans.Serialization/Codecs/WellKnownStringComparerCodec.cs
… src/Orleans.Serialization/ISerializableSerializer/DotNetSerializableCodec.cs
… src/Orleans.Serialization/ISerializableSerializer/ExceptionCodec.cs
… src/Orleans.Serialization/ISerializableSerializer/SerializationEntryCodec.cs
… src/Orleans.Serialization/ISerializableSerializer/ValueTypeSerializer.cs
… src/Orleans.Serialization/Utilities/VarIntWriterExtensions.cs
… test/Benchmarks/Serialization/ComplexTypeBenchmarks.cs