* Always use `SchemaType.Expected` encoding for static codecs (when used statically).
* Make `UriCodec` a static codec.
* Optimize integer and string serialization.

Some more complex types serialization is around 10% faster, but for `MyVector3` it's 1.35-1.4x faster.

|               main |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
|    OrleansPipeWriter | 50.46 μs | 0.264 μs | 0.234 μs |      - |     480 B |
|     OrleansSerialize | 25.99 μs | 0.057 μs | 0.053 μs | 0.4272 |   17032 B |
|  OrleansBufferWriter | 22.09 μs | 0.116 μs | 0.103 μs |      - |         - |
| OrleansBufferWriter2 | 49.57 μs | 0.540 μs | 0.505 μs |      - |         - |
| OrleansMessageSerializerStructRoundTrip | 1.096 μs | 0.0048 μs | 0.0040 μs | 0.0095 |     336 B |
|  OrleansMessageSerializerClassRoundTrip | 1.637 μs | 0.0227 μs | 0.0201 μs | 0.0114 |     448 B |

|               PR |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|--------------------- |---------:|---------:|---------:|-------:|----------:|
|    OrleansPipeWriter | 36.27 μs | 0.643 μs | 1.253 μs |      - |     480 B |
|     OrleansSerialize | 19.60 μs | 0.392 μs | 0.467 μs | 0.4272 |   17032 B |
|  OrleansBufferWriter | 16.67 μs | 0.263 μs | 0.246 μs |      - |         - |
| OrleansBufferWriter2 | 33.51 μs | 0.555 μs | 0.519 μs |      - |         - |
| OrleansMessageSerializerStructRoundTrip | 1.079 μs | 0.0088 μs | 0.0117 μs | 0.0095 |     336 B |
|  OrleansMessageSerializerClassRoundTrip | 1.555 μs | 0.0064 μs | 0.0060 μs | 0.0114 |     448 B |


###### Microsoft Reviewers: [Open in CodeFlow](https://portal.fabricbot.ms/api/codeflow?pullrequest=https://github.com/dotnet/orleans/pull/8206)